Update doi.md
Re-wrote the paragraphs for DOIs and the 3 rs levels. Many of the level-specific descriptors applied to all levels. As it was it either resulted in a suggestion of exclusion or resulted in repetition. (Note that the prev. version included some Dave edits, but this issue was also true pre-Dave edits.)